src/main/java/ru/netology/domain/TicketManager.java
package ru.netology.domain;

import java.util.Arrays;

public class TicketManager {
    private TicketRepository repository;

    public TicketManager(TicketRepository repository) {
        this.repository = repository;
    }

    public void add(Ticket ticket) {
        repository.save(ticket);
    }

    // метод определения соответствия товара product запросу search
    public boolean matches(Ticket ticket, String search) {
        if (ticket.getArrivalAirport().contains(search) || ticket.getDepartureAirport().contains(search)) {
            return true;
        } else {
            return false;
        }
    }

    public void removeById(int id) {
        repository.removeById(id);
    }

    public Ticket[] searchBy(String text) {
        Ticket[] result = new Ticket[0]; // тут будем хранить подошедшие запросу продукты
        for (Ticket ticket : repository.findAll()) {
            if (matches(ticket, text)) {
                Ticket[] tmp = new Ticket[result.length + 1];
                System.arraycopy(result, 0, tmp, 0, result.length);
                tmp[tmp.length - 1] = ticket;
                result = tmp;
            }
        }
        return result;
    }

    public Ticket[] findAll(String from, String to) {
        Ticket[] result = new Ticket[0];
        for (Ticket ticket : repository.findAll()) {
            if (matches(ticket, from) && matches(ticket, to)) {
                Ticket[] tmp = new Ticket[result.length + 1];
                System.arraycopy(result, 0, tmp, 0, result.length);
                tmp[tmp.length - 1] = ticket;
                result = tmp;
            }
        }
        Arrays.sort(result);
        return result;
    }

    public Ticket[] findAllTicket() {
        return repository.findAll();
    }
}
